# services/chat-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "실시간 채팅 서비스 API",
        "title": "Chat Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.wealist.co.kr/support",
            "email": "support@wealist.co.kr"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/api/chats",
    "paths": {
...
        "/presence/online": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "현재 온라인인 사용자 목록을 가져옵니다",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "presence"
                ],
                "summary": "온라인 사용자 목록",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/presence/status/{userId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "특정 사용자가 온라인인지 확인합니다",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "presence"
                ],
                "summary": "사용자 온라인 여부 확인",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
        "/ws/chat/{chatId}": {
            "get": {
                "description": "채팅방 WebSocket에 연결합니다",
                "tags": [
                    "websocket"
                ],
                "summary": "WebSocket 연결",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "chatId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT Access Token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}